+++ SDK/NIMSDK/ExportHeaders/Protocol/NIMChatManagerProtocol.h
@@ -50,7 +50,7 @@ typedef void(^NIMRevokeMessageBlock)(NSError * __nullable error);
  *  @param progress 进度
  */
 - (void)sendMessage:(NIMMessage *)message
-           progress:(CGFloat)progress;
+           progress:(float)progress;
 
 /**
  *  发送消息完成回调
@@ -95,7 +95,7 @@ typedef void(^NIMRevokeMessageBlock)(NSError * __nullable error);
  *  @discussion 附件包括:图片,视频的缩略图,语音文件
  */
 - (void)fetchMessageAttachment:(NIMMessage *)message
-                      progress:(CGFloat)progress;
+                      progress:(float)progress;
 
 
 /**
@@ -202,7 +202,7 @@ typedef void(^NIMRevokeMessageBlock)(NSError * __nullable error);
  *
  *  @return 正在传输的消息进度,如果消息不在传输,则返回0
  */
-- (CGFloat)messageTransportProgress:(NIMMessage *)message;
+- (float)messageTransportProgress:(NIMMessage *)message;
 
 /**
  *  添加聊天委托
